+++ utils/common.js
@@ -1,12 +1,18 @@
 const _ = require('lodash');
 const fs = require('fs');
-const crypto = require('crypto');
-const pem2jwk = require('pem-jwk').pem2jwk
-const jwt = require('jsonwebtoken');
+const { getJWTKeys } = require('./jwtHelper');
 
 var options = {};
 
 const getOptions = () => options;
+
+const _getValueFromString = strVal => {
+    try {
+        return JSON.parse(strVal);
+    } catch {
+        return strVal;
+    }
+};
 
 const getArgs = args => {
     const myArgs = {};
@@ -29,36 +35,6 @@ const getArgs = args => {
     return myArgs;
 };
 
-const getJWTKeys = async (privateKey, publicKey) => {
-    let effectivePrivateKey = getPEMString(privateKey);
-    let effectivePublicKey = getPEMString(publicKey);
-
-    if (effectivePrivateKey === null) {
-        const { privateKey, publicKey } = await getKeyPair();
-        effectivePrivateKey = privateKey;
-        effectivePublicKey = publicKey;
-    } else if (effectivePublicKey === null) {
-        const privateKey = crypto.createPrivateKey(effectivePrivateKey);
-        effectivePublicKey = privateKey.export({ type: 'spki', format: 'pem' });
-    }
-
-    return { privateKey: effectivePrivateKey, publicKey: effectivePublicKey }
-};
-
-const getPEMString = (key) => {
-    if (isValidPEM(key)) return key;
-    if (fs.existsSync(key)) {
-        const fileContent = fs.readFileSync(key, 'utf8');
-        if (isValidPEM(fileContent)) return fileContent;
-    }
-    return null;
-};
-
-function isValidPEM(pemString) {
-    const pemRegex = /^-----BEGIN [A-Z\s]+-----\r?\n[\/+=a-zA-Z0-9\r\n]*\r?\n-----END [A-Z\s]+-----\r?\n$/;
-    return pemRegex.test(pemString);
-}
-
 const getPrintableString = obj => {
     return _(obj)
         .keys()
@@ -72,48 +48,6 @@ const getPrintableString = obj => {
         .join('\r\n')
         .valueOf();
 };
-
-const getKeyPair = () => {
-    return new Promise((resolve, reject) => {
-        crypto.generateKeyPair('rsa', {
-            modulusLength: 4096,
-            publicKeyEncoding: {
-                type: 'spki',
-                format: 'pem'
-            },
-            privateKeyEncoding: {
-                type: 'pkcs8',
-                format: 'pem',
-            }
-        }, (err, publicKey, privateKey) => {
-            if (err) reject(err);
-            else resolve({ publicKey, privateKey });
-        });
-    });
-}
-
-const generateJWT = (user, privateKey, audience) => {
-    return jwt.sign(
-        user,
-        privateKey,
-        {
-            audience,
-            issuer: 'MockServer',
-            expiresIn: '1d',
-            algorithm: 'RS256',
-        }
-    );
-}
-
-function pemToJwk(pem) {
-    const jwk = {
-        ...pem2jwk(pem),
-        alg: 'RS256',
-    };
-    return jwk;
-}
-
-var options = {};
 
 const populateOptions = async () => {
     const args = getArgs(process.argv.slice(2));
@@ -134,4 +68,4 @@ const populateOptions = async () => {
 
 
 
-module.exports = { getOptions, populateOptions, getPrintableString, getKeyPair, generateJWT, pemToJwk };
+module.exports = { getOptions, populateOptions, getPrintableString };
